requirements.txt
asgiref==3.2.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-import-export==2.0.2
django-mdeditor==0.1.18
django-simpleui==3.9.1
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
mistune==0.8.4
odfpy==1.4.1
openpyxl==3.0.3
PyMySQL==0.9.3
pytz==2019.3
PyYAML==5.3.1
sqlparse==0.3.1
tablib==1.1.0
xlrd==1.2.0
xlwt==1.3.0
mysqlclient==1.3.14
gunicorn==19.9.0